Silence warning.

+++ filters/analyze.c
@@ -103,7 +103,7 @@ static inline int GetMagickNumberThreads(const Image *source,
        (GetImagePixelCacheType(source) != MapCache)) ||
       ((GetImagePixelCacheType(destination) != MemoryCache) &&
        (GetImagePixelCacheType(destination) != MapCache)))
-    return(MagickMax(MagickMin(GetMagickResourceLimit(ThreadResource),2),1));
+    return(MagickMax(MagickMin((int) GetMagickResourceLimit(ThreadResource),2),1));
   return(MagickMax(MagickMin((int) GetMagickResourceLimit(ThreadResource),
     (int) (chunk)/64),1));
 }
